Starting the refactoring of presenters into a common base attribute presenter
