+++ MvvmCross/Core/Core/Views/MvxViewPresenter.cs
@@ -1,4 +1,4 @@
-﻿// MvxViewPresenter.cs
+// MvxViewPresenter.cs
 
 // MvvmCross is licensed using Microsoft Public License (Ms-PL)
 // Contributions and inspirations noted in readme.md and license.txt
