+++ MvvmCross/iOS/iOS/Views/Presenters/MvxIosViewPresenter.cs
@@ -12,7 +12,7 @@ using MvvmCross.Platform;
 
 namespace MvvmCross.iOS.Views.Presenters
 {
-    public class MvxIosViewPresenter : MvxViewPresenter, IMvxIosViewPresenter, IMvxAttributeViewPresenter
+    public class MvxIosViewPresenter : MvxAttributeViewPresenter, IMvxIosViewPresenter
     {
         protected readonly IUIApplicationDelegate _applicationDelegate;
         protected readonly UIWindow _window;
@@ -24,77 +24,8 @@ namespace MvvmCross.iOS.Views.Presenters
         public IMvxTabBarViewController TabBarViewController { get; protected set; }
 
         public IMvxSplitViewController SplitViewController { get; protected set; }
-
-        private IMvxViewModelTypeFinder _viewModelTypeFinder;
-        public IMvxViewModelTypeFinder ViewModelTypeFinder
-        {
-            get
-            {
-                if (_viewModelTypeFinder == null)
-                    _viewModelTypeFinder = Mvx.Resolve<IMvxViewModelTypeFinder>();
-                return _viewModelTypeFinder;
-            }
-            set
-            {
-                _viewModelTypeFinder = value;
-            }
-        }
-
-        private IMvxViewsContainer _viewsContainer;
-        public IMvxViewsContainer ViewsContainer
-        {
-            get
-            {
-                if (_viewsContainer == null)
-                    _viewsContainer = Mvx.Resolve<IMvxViewsContainer>();
-                return _viewsContainer;
-            }
-            set
-            {
-                _viewsContainer = value;
-            }
-        }
-
-        private Dictionary<Type, MvxPresentationAttributeAction> _attributeTypesActionsDictionary;
-        public Dictionary<Type, MvxPresentationAttributeAction> AttributeTypesToActionsDictionary
-        {
-            get
-            {
-                if (_attributeTypesActionsDictionary == null)
-                {
-                    _attributeTypesActionsDictionary = new Dictionary<Type, MvxPresentationAttributeAction>();
-                    RegisterAttributeTypes();
-                }
-                return _attributeTypesActionsDictionary;
-            }
-        }
-
-        public virtual MvxBasePresentationAttribute GetPresentationAttribute(Type viewModelType)
-        {
-            var viewType = ViewsContainer.GetViewType(viewModelType);
-
-            var overrideAttribute = GetOverridePresentationAttribute(viewModelType, viewType);
-            if (overrideAttribute != null)
-                return overrideAttribute;
-
-            var attribute = viewType
-                .GetCustomAttributes(typeof(MvxBasePresentationAttribute), true)
-                .FirstOrDefault() as MvxBasePresentationAttribute;
-            if (attribute != null)
-            {
-                if (attribute.ViewType == null)
-                    attribute.ViewType = viewType;
-
-                if (attribute.ViewModelType == null)
-                    attribute.ViewModelType = viewModelType;
-
-                return attribute;
-            }
-
-            return CreatePresentationAttribute(viewModelType, viewType);
-        }
-
-        public virtual MvxBasePresentationAttribute CreatePresentationAttribute(Type viewModelType, Type viewType)
+        
+        public override MvxBasePresentationAttribute CreatePresentationAttribute(Type viewModelType, Type viewType)
         {
             if (MasterNavigationController == null &&
                (TabBarViewController == null ||
@@ -110,10 +41,11 @@ namespace MvvmCross.iOS.Views.Presenters
             return new MvxChildPresentationAttribute() { ViewType = viewType, ViewModelType = viewModelType };
         }
 
-        public virtual MvxBasePresentationAttribute GetOverridePresentationAttribute(Type viewModelType, Type viewType)
+        public override MvxBasePresentationAttribute GetOverridePresentationAttribute(Type viewModelType, Type viewType)
         {
             if (viewType?.GetInterface(nameof(IMvxOverridePresentationAttribute)) != null)
             {
+
                 var viewInstance = this.CreateViewControllerFor(viewType, null) as UIViewController;
                 using (viewInstance)
                 {
@@ -145,7 +77,7 @@ namespace MvvmCross.iOS.Views.Presenters
             _window = window;
         }
 
-        public virtual void RegisterAttributeTypes()
+        public override void RegisterAttributeTypes()
         {
             AttributeTypesToActionsDictionary.Add(
                 typeof(MvxRootPresentationAttribute),
