+++ TestProjects/Forms/PageRenderer/PageRenderer.iOS/PageRendererExample.iOS.csproj
@@ -128,7 +128,7 @@
   </ItemGroup>
   <ItemGroup>
     <ImageAsset Include="Resources\Images.xcassets\AppIcons.appiconset\Contents.json">
-      <InProject>false</InProject>
+      <Visible>false</Visible>
     </ImageAsset>
   </ItemGroup>
   <ItemGroup>
